WEB-1986 refactor logo
