+++ src/logo-movistar.tsx
@@ -1,0 +1,70 @@
+import * as React from 'react';
+import * as styles from './logo.css';
+import {getMovistarSkin} from './skins/movistar';
+import {applyCssVars} from './utils/css';
+import {calcInlineVars} from './logo-common';
+
+import type {LogoImageProps} from './logo-common';
+
+const MovistarLogoImage = ({size, type, isDarkMode, isInverse}: LogoImageProps): JSX.Element => {
+    const {colors} = getMovistarSkin();
+    const color = isInverse && !isDarkMode ? colors.inverse : colors.brand;
+
+    if (type === 'vertical') {
+        return (
+            <svg
+                className={styles.svg}
+                style={applyCssVars(calcInlineVars(size))}
+                viewBox="0 0 72 72"
+                fill={color}
+                role="presentation"
+            >
+                <path d="M22.1005 12.8312C20.077 12.8698 16.3384 13.8526 14.6232 20.8289C13.8716 23.8738 13.5825 27.0343 14.2378 30.8115C14.8159 34.2804 15.8566 37.2867 16.5696 38.9441C16.8201 39.5222 17.1863 40.1004 17.4754 40.4665C18.304 41.5265 19.6916 41.4494 20.2697 41.1603C20.9057 40.852 21.638 40.1004 21.3682 38.366C21.2333 37.5373 20.8672 36.3232 20.6744 35.6487C20.0385 33.5674 19.1905 31.0621 19.1134 29.2891C19.0171 26.8994 19.9614 26.5911 20.5781 26.4562C21.638 26.2249 22.5052 27.3812 23.3532 28.8266C24.3553 30.561 26.0705 33.6059 27.458 35.9377C28.7107 38.0383 31.0425 40.3124 34.7812 40.1582C38.5969 40.004 41.3913 38.5394 42.8367 33.9721C43.9159 30.5417 44.6482 27.9979 45.843 25.377C47.2113 22.3706 49.0228 20.7518 50.5645 21.2529C51.9906 21.7154 52.3375 23.1029 52.3568 25.1457C52.376 26.9572 52.1641 28.9422 51.9906 30.4068C51.9328 30.9464 51.8172 32.0064 51.875 32.6038C51.9521 33.7793 52.4724 34.9356 53.7829 35.1283C55.1897 35.3211 56.3074 34.2033 56.7507 32.8543C56.9241 32.3147 57.0783 31.5053 57.1554 30.9272C57.5601 27.9979 57.6757 26.0322 57.483 23.0451C57.271 19.557 56.5772 16.3579 55.3824 13.6021C54.2454 10.9619 52.4146 9.266 50.0442 9.11183C47.4425 8.93839 44.4555 10.6728 42.8752 14.0261C41.4298 17.1095 40.2736 20.2893 39.5798 21.9081C38.8667 23.5462 37.8261 24.5676 36.2073 24.741C34.2416 24.953 32.5457 23.5269 31.3123 21.4841C30.2331 19.7112 28.094 16.3386 26.9377 15.2016C25.897 14.1224 24.6636 12.7927 22.1005 12.8312Z" />
+                <path d="M17.3019 51.9718V62.494H15.8758C15.5482 62.494 15.3748 62.3013 15.3748 61.9737V55.0167L12.5419 59.6611H11.3085C10.9809 59.6611 10.8267 59.5455 10.6725 59.2949L8.01307 55.1516V62.494H6.06665V52.5692C6.06665 52.1838 6.25937 51.9718 6.68334 51.9718H8.24433L11.771 57.6183L14.8737 52.415C15.0664 52.0681 15.2592 51.9718 15.6639 51.9718H17.3019Z" />
+                <path d="M18.8244 58.7938C18.8244 56.5198 20.3468 54.9395 22.6594 54.9395C24.972 54.9395 26.4944 56.5198 26.4944 58.7938C26.4944 61.0679 24.9527 62.6481 22.6594 62.6481C20.3468 62.6481 18.8244 61.0679 18.8244 58.7938ZM24.3938 58.7938C24.3938 57.4063 23.7579 56.5391 22.6594 56.5391C21.5609 56.5391 20.925 57.387 20.925 58.7938C20.925 60.1621 21.5609 61.0486 22.6594 61.0486C23.7579 61.0486 24.3938 60.1814 24.3938 58.7938Z" />
+                <path d="M35.1473 55.0938L31.9868 62.4748H30.5414C30.2331 62.4748 30.0597 62.3591 29.9633 62.0701L27.246 55.6719C27.1111 55.325 27.246 55.0745 27.6315 55.0745H29.231L31.1196 60.2007L33.0082 55.0745H35.1473V55.0938Z" />
+                <path d="M35.9567 52.5499C35.9567 51.8754 36.5156 51.3358 37.1901 51.3358C37.8453 51.3358 38.4042 51.8754 38.4042 52.5499C38.4042 53.2629 37.8453 53.7832 37.1901 53.7832C36.4963 53.7832 35.9567 53.2629 35.9567 52.5499ZM36.1302 55.614C36.1302 55.2671 36.3229 55.0937 36.6505 55.0937H38.1922V62.494H36.1302V55.614Z" />
+                <path d="M40.3699 62.3976C39.9459 62.282 39.8688 62.0892 39.8688 61.8194V60.7788C40.6204 60.9715 41.5455 61.1449 42.3741 61.1449C43.4148 61.1449 43.8195 60.7981 43.8195 60.3548C43.8195 59.873 43.3763 59.7381 42.6825 59.5647L41.9502 59.372C40.5819 59.0443 39.7147 58.3891 39.7147 57.1365C39.7147 55.6526 40.8902 54.9395 42.9523 54.9395C43.6846 54.9395 44.4362 55.0552 44.8602 55.1901C45.2649 55.3057 45.3227 55.4599 45.3227 55.7682V56.7318C44.5904 56.5391 43.6846 56.4427 42.9716 56.4427C42.0465 56.4427 41.6611 56.674 41.6611 57.0787C41.6611 57.4834 42.1814 57.6761 42.8367 57.8495L43.6653 58.0615C44.706 58.312 45.7852 58.6975 45.7852 60.2585C45.7852 61.858 44.5711 62.6674 42.432 62.6674C41.6804 62.6481 40.8324 62.5518 40.3699 62.3976Z" />
+                <path d="M47.6353 60.2778V56.5584H46.1128V55.6141C46.1128 55.2672 46.3055 55.0938 46.6331 55.0938H47.6353V52.9546C47.6353 52.627 47.828 52.4343 48.1556 52.4343H49.6973V55.0938H51.663V56.5584H49.6973V59.9117C49.6973 60.6825 49.9671 61.0679 50.8536 61.0679C51.2583 61.0679 51.7015 60.9909 52.0677 60.9138V61.9159C52.0677 62.1471 52.0099 62.3013 51.6437 62.4555C51.3547 62.5711 50.8343 62.6482 50.2947 62.6482C48.6952 62.6482 47.6353 61.8195 47.6353 60.2778Z" />
+                <path d="M60.7977 61.0294V61.993C60.7977 62.1857 60.7977 62.3977 60.4315 62.5133C60.2388 62.5711 59.9497 62.6289 59.6414 62.6289C58.6393 62.6289 58.0997 62.1664 57.8491 61.4148C57.3095 62.1857 56.4231 62.6482 55.4017 62.6482C53.8792 62.6482 52.8771 61.6075 52.8771 60.2778C52.8771 58.7168 54.2454 57.9267 56.9048 57.9267C57.1746 57.9267 57.4637 57.946 57.695 57.9652C57.695 56.9631 57.2325 56.5584 55.922 56.5584C55.1897 56.5584 54.2261 56.674 53.5131 56.8668V55.8646C53.5131 55.537 53.5902 55.3829 53.9563 55.2672C54.4766 55.0938 55.3631 54.9782 56.2303 54.9782C58.9476 54.9782 59.757 56.1923 59.757 57.7532V60.3742C59.757 60.9138 59.8727 61.145 60.3544 61.145C60.4893 61.0872 60.7013 61.0679 60.7977 61.0294ZM57.6564 59.6611V59.1215H57.271C55.6715 59.1215 54.8621 59.3721 54.8621 60.2007C54.8621 60.7596 55.2668 61.2221 56.0184 61.2221C56.9434 61.2221 57.6564 60.5862 57.6564 59.6611Z" />
+                <path d="M66.7333 55.4213V56.8474H66.0203C64.7098 56.8474 64.016 57.3677 64.016 58.8709V62.5132H61.954V55.6333C61.954 55.2864 62.1467 55.1129 62.455 55.1129H63.8426L63.8811 56.6739C64.2858 55.5562 65.1916 54.9588 66.2323 54.9588C66.5599 54.9395 66.7333 55.0937 66.7333 55.4213Z" />
+            </svg>
+        );
+    }
+
+    if (type === 'imagotype') {
+        return (
+            <svg
+                className={styles.svg}
+                style={applyCssVars(calcInlineVars(size))}
+                viewBox="0 0 195 72"
+                fill={color}
+                role="presentation"
+            >
+                <path d="M17.9412 17.6705C15.0439 17.7257 9.69094 19.1342 7.23519 29.1317C6.15907 33.4952 5.74518 38.0245 6.68333 43.4375C7.51111 48.4086 9.00112 52.7169 10.022 55.092C10.3808 55.9205 10.905 56.7491 11.3189 57.2738C12.5054 58.7927 14.4921 58.6823 15.3198 58.268C16.2304 57.8261 17.2789 56.7491 16.8926 54.2635C16.6995 53.0759 16.1752 51.3361 15.8993 50.3694C14.9887 47.3868 13.7747 43.7965 13.6643 41.2557C13.5263 37.8312 14.8784 37.3893 15.7613 37.196C17.2789 36.8646 18.5206 38.5216 19.7347 40.5929C21.1695 43.0785 23.6253 47.442 25.6119 50.7837C27.4055 53.794 30.7442 57.0528 36.0972 56.8319C41.5605 56.611 45.5615 54.5121 47.6309 47.9667C49.1761 43.0509 50.2246 39.4054 51.9354 35.6494C53.8944 31.3411 56.4882 29.0212 58.6956 29.7393C60.7374 30.4021 61.2341 32.3906 61.2617 35.318C61.2893 37.914 60.9858 40.7586 60.7374 42.8575C60.6547 43.6308 60.4891 45.1498 60.5719 46.0059C60.6823 47.6906 61.4273 49.3476 63.3036 49.6238C65.3178 49.9 66.9182 48.2982 67.5528 46.3649C67.8012 45.5916 68.0219 44.4317 68.1323 43.6032C68.7117 39.4054 68.8773 36.5884 68.6014 32.3077C68.2978 27.309 67.3045 22.7245 65.5938 18.7752C63.9658 14.9916 61.3445 12.5613 57.9506 12.3404C54.2256 12.0918 49.9487 14.5774 47.6861 19.3828C45.6166 23.8016 43.9611 28.3584 42.9677 30.6783C41.9468 33.0258 40.4568 34.4895 38.139 34.738C35.3246 35.0418 32.8964 32.9981 31.1305 30.0707C29.5853 27.5299 26.5225 22.6969 24.8669 21.0674C23.3493 19.5209 21.611 17.6153 17.9412 17.6705Z" />
+                <path d="M107.038 44.5422H104.527C103.948 44.5422 103.644 44.2384 103.644 43.6308V31.3687L98.6498 39.5434H96.47C96.1941 39.5434 95.9733 39.4882 95.8078 39.4054C95.6422 39.2949 95.5042 39.1292 95.3663 38.9082L90.7031 31.5896V44.5422H87.2816V27.0328C87.2816 26.3423 87.6403 25.9833 88.3577 25.9833H91.117L97.353 35.9532L102.816 26.7842C102.982 26.4804 103.175 26.2595 103.368 26.1766C103.561 26.0662 103.837 26.0109 104.196 26.0109H107.038V44.5422Z" />
+                <path d="M112.943 43.9623C111.922 43.3823 111.149 42.5814 110.57 41.5595C109.99 40.5377 109.714 39.3502 109.714 38.0245C109.714 36.6989 109.99 35.5113 110.57 34.4895C111.149 33.4677 111.922 32.6668 112.943 32.0868C113.964 31.5068 115.15 31.2307 116.502 31.2307C117.854 31.2307 119.041 31.5068 120.062 32.0868C121.083 32.6668 121.855 33.4677 122.435 34.4895C123.014 35.5113 123.29 36.6989 123.29 38.0245C123.29 39.3502 123.014 40.5377 122.435 41.5595C121.855 42.5814 121.083 43.3823 120.062 43.9623C119.041 44.5422 117.854 44.8184 116.53 44.8184C115.15 44.8184 113.964 44.5422 112.943 43.9623ZM118.71 40.9243C119.262 40.2063 119.51 39.2397 119.51 38.0245C119.51 36.7817 119.234 35.8151 118.71 35.0971C118.185 34.4067 117.44 34.0476 116.447 34.0476C115.481 34.0476 114.736 34.4067 114.185 35.0971C113.66 35.7875 113.384 36.7817 113.384 38.0245C113.384 39.2397 113.66 40.2063 114.185 40.9243C114.736 41.6424 115.481 42.0014 116.42 42.0014C117.44 42.0014 118.185 41.6424 118.71 40.9243Z" />
+                <path d="M134.686 31.5068H138.466L132.92 44.5422H130.381C130.106 44.5422 129.885 44.487 129.747 44.3765C129.581 44.266 129.471 44.1003 129.36 43.8242L124.587 32.5563C124.532 32.3906 124.504 32.2525 124.504 32.1144C124.504 31.9211 124.559 31.7554 124.697 31.6449C124.835 31.5344 125.001 31.4792 125.277 31.4792H128.091L131.402 40.5101L134.686 31.5068Z" />
+                <path d="M140.977 28.8831C140.646 28.6898 140.398 28.4413 140.204 28.1099C140.011 27.7785 139.901 27.4194 139.901 27.0052C139.901 26.6185 140.011 26.2595 140.204 25.9281C140.398 25.5967 140.674 25.3481 140.977 25.1548C141.308 24.9615 141.667 24.8786 142.053 24.8786C142.439 24.8786 142.798 24.9615 143.129 25.1548C143.46 25.3481 143.736 25.5967 143.902 25.9281C144.095 26.2595 144.178 26.6185 144.178 27.0052C144.178 27.3918 144.095 27.7785 143.902 28.1099C143.709 28.4413 143.46 28.7174 143.129 28.8831C142.798 29.0765 142.439 29.1593 142.053 29.1593C141.667 29.1593 141.281 29.0765 140.977 28.8831ZM140.204 32.4182C140.204 31.8106 140.508 31.5068 141.115 31.5068H143.847V44.5698H140.204V32.4182Z" />
+                <path d="M149.282 44.7079C148.62 44.6251 148.068 44.5146 147.654 44.3765C147.296 44.2661 147.047 44.128 146.937 43.9899C146.827 43.8242 146.771 43.6308 146.771 43.3547V41.5319C147.489 41.7253 148.234 41.891 149.007 42.0014C149.779 42.1119 150.496 42.1671 151.186 42.1671C152.042 42.1671 152.676 42.029 153.09 41.7805C153.504 41.5319 153.725 41.1729 153.725 40.7586C153.725 40.4825 153.642 40.2615 153.504 40.1234C153.339 39.9577 153.145 39.8197 152.869 39.7368C152.594 39.6263 152.235 39.5159 151.738 39.4054L150.441 39.0464C149.172 38.7426 148.206 38.2731 147.516 37.6379C146.827 37.0027 146.496 36.1742 146.496 35.0971C146.496 33.7991 146.992 32.8325 147.958 32.1973C148.924 31.5621 150.359 31.2307 152.207 31.2307C152.842 31.2307 153.449 31.2583 154.056 31.3411C154.663 31.424 155.16 31.5344 155.574 31.6449C155.905 31.7278 156.125 31.8382 156.236 32.0039C156.346 32.1696 156.374 32.363 156.374 32.6668V34.379C155.739 34.2133 155.022 34.1029 154.304 33.9924C153.559 33.9095 152.869 33.8543 152.235 33.8543C150.717 33.8543 149.945 34.2133 149.945 34.959C149.945 35.2904 150.11 35.5666 150.469 35.7599C150.828 35.9532 151.324 36.1465 152.014 36.3122L153.477 36.6989C154.222 36.8922 154.856 37.0855 155.38 37.3341C155.905 37.5827 156.346 37.9693 156.677 38.494C157.008 39.0188 157.202 39.6816 157.202 40.5653C157.202 41.9738 156.705 43.0233 155.684 43.7413C154.663 44.4594 153.201 44.8184 151.297 44.8184C150.634 44.8184 149.972 44.7908 149.282 44.7079Z" />
+                <path d="M162.665 44.3489C161.975 44.0175 161.423 43.548 161.037 42.9128C160.651 42.2776 160.457 41.5319 160.457 40.6482V34.1028H157.781V32.4182C157.781 31.8106 158.084 31.5068 158.692 31.5068H160.457V27.7232C160.457 27.4194 160.54 27.1985 160.706 27.0328C160.871 26.8671 161.092 26.8119 161.368 26.8119H164.1V31.5068H167.576V34.1028H164.1V40.013C164.1 40.7034 164.238 41.2281 164.569 41.5595C164.872 41.8909 165.397 42.0566 166.142 42.0566C166.721 42.0566 167.438 41.9738 168.266 41.7805V43.548C168.266 43.7689 168.211 43.9346 168.128 44.1003C168.018 44.2384 167.825 44.3765 167.521 44.5146C167.273 44.6251 166.914 44.7079 166.5 44.7631C166.059 44.8184 165.617 44.846 165.148 44.846C164.182 44.8184 163.382 44.6527 162.665 44.3489Z" />
+                <path d="M183.249 42.0566C183.387 42.029 183.525 42.0014 183.635 41.9738V43.6861C183.635 43.907 183.608 44.1003 183.525 44.2384C183.442 44.4041 183.277 44.5146 183.001 44.5974C182.587 44.7355 182.118 44.7908 181.593 44.7908C179.965 44.7908 178.917 44.0727 178.448 42.6642C177.979 43.3547 177.372 43.8794 176.627 44.266C175.882 44.6527 175.054 44.846 174.143 44.846C173.288 44.846 172.515 44.6527 171.826 44.2936C171.136 43.9346 170.612 43.4099 170.253 42.7747C169.894 42.1395 169.701 41.4214 169.701 40.6482C169.701 39.2673 170.308 38.2454 171.495 37.555C172.681 36.8646 174.447 36.5056 176.792 36.5056L178.172 36.5332C178.144 35.9256 178.062 35.4561 177.868 35.1247C177.675 34.7657 177.372 34.5171 176.903 34.3514C176.434 34.1857 175.854 34.1028 175.082 34.1028C174.447 34.1028 173.73 34.1581 172.929 34.2409C172.157 34.3514 171.439 34.4619 170.805 34.6276V32.8601C170.805 32.5563 170.86 32.3353 170.97 32.1696C171.081 32.0039 171.274 31.8935 171.605 31.8106C172.074 31.6725 172.681 31.5344 173.398 31.4516C174.116 31.3411 174.861 31.3135 175.606 31.3135C177.841 31.3135 179.441 31.7554 180.407 32.6115C181.373 33.4953 181.842 34.6828 181.842 36.2018V40.8139C181.842 41.311 181.925 41.6424 182.063 41.8633C182.201 42.0843 182.476 42.1671 182.89 42.1671C182.973 42.0843 183.111 42.0566 183.249 42.0566ZM176.71 41.9738C177.151 41.7528 177.482 41.4214 177.758 40.9796C178.006 40.5653 178.144 40.0682 178.144 39.5435V38.6045H177.482C176.02 38.6045 174.944 38.7426 174.254 39.0463C173.564 39.3225 173.205 39.8196 173.205 40.5377C173.205 41.0624 173.398 41.5043 173.757 41.8357C174.116 42.1671 174.613 42.3604 175.275 42.3604C175.771 42.3052 176.268 42.1947 176.71 41.9738Z" />
+                <path d="M194.093 32.0592V34.5723H192.824C191.61 34.5723 190.727 34.8485 190.147 35.4009C189.568 35.9532 189.292 36.8646 189.292 38.135V44.5698H185.65V32.4182C185.65 31.8106 185.953 31.5068 186.533 31.5068H188.988L189.044 34.2685C189.402 33.3019 189.954 32.5563 190.671 32.0315C191.389 31.5068 192.244 31.2583 193.155 31.2583C193.789 31.2306 194.093 31.5068 194.093 32.0592Z" />
+            </svg>
+        );
+    }
+
+    return (
+        <svg
+            className={styles.svg}
+            style={applyCssVars(calcInlineVars(size))}
+            viewBox="0 0 72 72"
+            fill={color}
+            role="presentation"
+        >
+            <path d="M17.2797 17.2124C14.4995 17.2572 9.36782 18.6502 7.02384 28.3781C6.00256 32.6168 5.60793 37.0347 6.4823 42.2938C7.28828 47.1441 8.71635 51.3303 9.67886 53.6376C10.0107 54.4324 10.5244 55.2617 10.9215 55.7719C12.0639 57.239 13.9651 57.1453 14.764 56.7453C15.6348 56.3098 16.6358 55.2581 16.2736 52.8561C16.0983 51.6958 15.5933 49.9981 15.3091 49.0545C14.4367 46.1588 13.2756 42.6649 13.1743 40.176C13.039 36.8468 14.3308 36.4108 15.1885 36.2193C16.6312 35.8965 17.8405 37.5063 18.9894 39.5237C20.3607 41.9314 22.7108 46.1974 24.6277 49.456C26.3582 52.3975 29.5517 55.5469 34.6804 55.3307C39.9089 55.1109 43.763 53.0837 45.7483 46.7014C47.2341 41.9267 48.2468 38.3588 49.8765 34.7048C51.7503 30.5031 54.2498 28.254 56.3547 28.9408C58.3091 29.5781 58.7969 31.5193 58.8202 34.3728C58.841 36.8968 58.5538 39.6792 58.3304 41.7234C58.2493 42.4652 58.1024 43.9576 58.1622 44.7859C58.2802 46.4146 58.9742 48.0414 60.7792 48.3018C62.7017 48.5788 64.2443 47.0174 64.8603 45.1302C65.1034 44.3864 65.3106 43.2482 65.4226 42.4405C65.9874 38.3618 66.1338 35.6206 65.8795 31.4473C65.5811 26.5676 64.6414 22.1178 63.0021 18.2667C61.4337 14.5839 58.9145 12.2236 55.6835 12.0151C52.107 11.7845 48.0026 14.1958 45.8496 18.8741C43.8648 23.1865 42.2772 27.6136 41.3141 29.873C40.3374 32.1643 38.9018 33.5764 36.6946 33.8127C33.996 34.101 31.6708 32.1103 29.9676 29.2718C28.4828 26.7977 25.5401 22.0869 23.9656 20.5039C22.4864 19.0167 20.7969 17.1558 17.2797 17.2124Z" />
+        </svg>
+    );
+};
+
+export default MovistarLogoImage;
