+++ src/logo-telefonica.tsx
@@ -1,0 +1,102 @@
+import * as React from 'react';
+import * as styles from './logo.css';
+import {getTelefonicaSkin} from './skins/telefonica';
+import {applyCssVars} from './utils/css';
+import {calcInlineVars} from './logo-common';
+
+import type {LogoImageProps} from './logo-common';
+
+const TelefonicaLogoImage = ({size, type, isDarkMode, isInverse}: LogoImageProps): JSX.Element => {
+    const {colors} = getTelefonicaSkin();
+    const color = isInverse && !isDarkMode ? colors.inverse : colors.brand;
+
+    if (type === 'vertical') {
+        return (
+            <svg
+                className={styles.svg}
+                style={applyCssVars(calcInlineVars(size))}
+                viewBox="0 0 73 72"
+                fill={color}
+                role="presentation"
+            >
+                <path d="M20.6921 19.2599C24.3945 19.2599 27.3958 16.2916 27.3958 12.63C27.3958 8.96833 24.3945 6 20.6921 6C16.9898 6 13.9885 8.96833 13.9885 12.63C13.9885 16.2916 16.9898 19.2599 20.6921 19.2599Z" />
+                <path d="M36.4448 19.2599C40.1471 19.2599 43.1484 16.2916 43.1484 12.63C43.1484 8.96833 40.1471 6 36.4448 6C32.7424 6 29.7411 8.96833 29.7411 12.63C29.7411 16.2916 32.7424 19.2599 36.4448 19.2599Z" />
+                <path d="M52.1973 19.2599C55.8997 19.2599 58.901 16.2916 58.901 12.63C58.901 8.96833 55.8997 6 52.1973 6C48.495 6 45.4937 8.96833 45.4937 12.63C45.4937 16.2916 48.495 19.2599 52.1973 19.2599Z" />
+                <path d="M36.4448 34.8394C40.1471 34.8394 43.1484 31.8711 43.1484 28.2095C43.1484 24.5479 40.1471 21.5795 36.4448 21.5795C32.7424 21.5795 29.7411 24.5479 29.7411 28.2095C29.7411 31.8711 32.7424 34.8394 36.4448 34.8394Z" />
+                <path d="M36.4448 50.4201C40.1471 50.4201 43.1484 47.4518 43.1484 43.7902C43.1484 40.1286 40.1471 37.1602 36.4448 37.1602C32.7424 37.1602 29.7411 40.1286 29.7411 43.7902C29.7411 47.4518 32.7424 50.4201 36.4448 50.4201Z" />
+                <path d="M16.0791 60.2396H13.9885V59.0084H19.4643V60.2396H17.3737V65.9012H16.0791V60.2396Z" />
+                <path d="M23.8347 64.424C23.7255 64.8176 23.0982 65.9996 21.4954 65.9996C20.0018 65.9996 18.9062 64.9196 18.9062 63.3896C18.9062 61.8596 19.9981 60.7808 21.4954 60.7808C22.8883 60.7808 23.9839 61.8608 23.9839 63.2912C23.9821 63.4118 23.972 63.532 23.9536 63.6512L23.9342 63.7892H20.151C20.2602 64.4876 20.788 64.97 21.4954 64.97C22.0826 64.97 22.4406 64.6352 22.5401 64.4288L23.8347 64.424ZM22.7427 62.9468C22.6335 62.3072 22.2052 61.814 21.499 61.814C20.7419 61.814 20.3039 62.3072 20.1547 62.9468H22.7427Z" />
+                <path d="M24.7301 59.0084H25.975V65.9012H24.7301V59.0084Z" />
+                <path d="M31.6485 64.424C31.5393 64.8176 30.912 65.9996 29.3092 65.9996C27.8156 65.9996 26.7212 64.9196 26.7212 63.3896C26.7212 61.8596 27.8132 60.7808 29.3092 60.7808C30.7033 60.7808 31.7989 61.8608 31.7989 63.2912C31.7965 63.4117 31.7864 63.532 31.7686 63.6512L31.748 63.7892H27.9648C28.074 64.4876 28.6018 64.97 29.3092 64.97C29.8964 64.97 30.2556 64.6352 30.3551 64.4288L31.6485 64.424ZM30.5565 62.9468C30.4461 62.3072 30.019 61.814 29.3116 61.814C28.5545 61.814 28.1165 62.3072 27.9672 62.9468H30.5565Z" />
+                <path d="M32.9431 62.0612H32.0938V60.8792H32.9431V60.14C32.9431 59.3924 33.4284 58.91 34.188 58.91H35.2799V59.99H34.5338C34.4876 59.9888 34.4417 59.997 34.3988 60.014C34.356 60.0309 34.3171 60.0564 34.2845 60.0887C34.2519 60.121 34.2263 60.1596 34.2093 60.2021C34.1923 60.2445 34.1842 60.2899 34.1855 60.3356V60.8768H35.2775V62.0588H34.1855V65.8987H32.9431V62.0612Z" />
+                <path d="M40.6465 63.3896C40.6465 64.916 39.5545 65.9996 38.0572 65.9996C36.56 65.9996 35.4741 64.9196 35.4741 63.3896C35.4741 61.8596 36.5661 60.7808 38.0621 60.7808C39.5581 60.7808 40.6465 61.8644 40.6465 63.3896ZM39.4016 63.3896C39.4016 62.504 38.8046 61.9136 38.0572 61.9136C37.3098 61.9136 36.7129 62.504 36.7129 63.3896C36.7129 64.2752 37.311 64.8668 38.0572 64.8668C38.8034 64.8668 39.4016 64.2764 39.4016 63.3896Z" />
+                <path d="M41.3927 60.8792H42.5369L42.6364 61.3712H42.6861C42.7852 61.2584 42.899 61.1591 43.0246 61.076C43.3396 60.8774 43.7067 60.7747 44.0802 60.7808C45.2256 60.7808 46.0713 61.6208 46.0713 62.8976V65.8976H44.8276V62.9996C44.8276 62.7122 44.7122 62.4366 44.5067 62.2334C44.3013 62.0302 44.0226 61.916 43.732 61.916C43.4414 61.916 43.1627 62.0302 42.9573 62.2334C42.7518 62.4366 42.6364 62.7122 42.6364 62.9996V65.9048H41.3927V60.8792Z" />
+                <path d="M53.9931 64.0304C53.8341 64.8176 53.2068 65.9996 51.604 65.9996C50.1104 65.9996 49.0148 64.9196 49.0148 63.3896C49.0148 61.8596 50.1068 60.7808 51.604 60.7808C53.2068 60.7808 53.8341 61.9628 53.9931 62.7008H52.7482C52.6697 62.4646 52.5156 62.2601 52.3092 62.1181C52.1028 61.9761 51.8553 61.9043 51.604 61.9136C50.8566 61.9136 50.2597 62.504 50.2597 63.3896C50.2597 64.2752 50.8566 64.8668 51.604 64.8668C51.8634 64.8785 52.1189 64.8013 52.3274 64.6483C52.5358 64.4953 52.6845 64.2757 52.7482 64.0268L53.9931 64.0304Z" />
+                <path d="M57.7568 65.4092H57.7071C57.6066 65.5247 57.4887 65.6243 57.3576 65.7044C57.03 65.9048 56.6508 66.0073 56.2656 65.9996C55.1106 65.9996 54.4238 65.3 54.4238 64.4732C54.4238 63.4892 55.1203 62.7992 56.5144 62.7992H57.6598V62.7008C57.6598 62.1488 57.3212 61.7648 56.7631 61.7648C56.205 61.7648 55.9138 62.1248 55.8677 62.4056H54.6204C54.7296 61.5788 55.4066 60.7808 56.7607 60.7808C58.0553 60.7808 58.901 61.628 58.901 62.7008V65.9012H57.8563L57.7568 65.4092ZM57.6573 63.7292H56.6114C55.9647 63.7292 55.6651 63.9692 55.6651 64.3688C55.6651 64.7684 55.9538 65.0084 56.4622 65.0084C57.2181 65.0084 57.6573 64.5752 57.6573 63.8768V63.7292Z" />
+                <path d="M47.0249 60.8792H48.2686V65.9012H47.0249V60.8792Z" />
+                <path d="M47.6474 60.1569C48.0542 60.1569 48.3839 59.8308 48.3839 59.4285C48.3839 59.0262 48.0542 58.7001 47.6474 58.7001C47.2407 58.7001 46.9109 59.0262 46.9109 59.4285C46.9109 59.8308 47.2407 60.1569 47.6474 60.1569Z" />
+                <path d="M38.5062 58.7624H39.7996L38.6045 60.1904H37.5598L38.5062 58.7624Z" />
+            </svg>
+        );
+    }
+
+    if (type === 'imagotype') {
+        return (
+            <svg
+                className={styles.svg}
+                style={applyCssVars(calcInlineVars(size))}
+                viewBox="0 0 195 72"
+                fill={color}
+                role="presentation"
+            >
+                <path d="M12.5551 28.1101C16.1753 28.1101 19.1101 25.1753 19.1101 21.555C19.1101 17.9348 16.1753 15 12.5551 15C8.9348 15 6 17.9348 6 21.555C6 25.1753 8.9348 28.1101 12.5551 28.1101Z" />
+                <path d="M27.9582 28.1101C31.5784 28.1101 34.5132 25.1753 34.5132 21.555C34.5132 17.9347 31.5784 14.9999 27.9582 14.9999C24.3379 14.9999 21.4031 17.9347 21.4031 21.555C21.4031 25.1753 24.3379 28.1101 27.9582 28.1101Z" />
+                <path d="M43.3631 28.1101C46.9834 28.1101 49.9182 25.1753 49.9182 21.555C49.9182 17.9347 46.9834 14.9999 43.3631 14.9999C39.7428 14.9999 36.808 17.9347 36.808 21.555C36.808 25.1753 39.7428 28.1101 43.3631 28.1101Z" />
+                <path d="M27.9582 43.515C31.5784 43.515 34.5132 40.5802 34.5132 36.96C34.5132 33.3397 31.5784 30.4049 27.9582 30.4049C24.3379 30.4049 21.4031 33.3397 21.4031 36.96C21.4031 40.5802 24.3379 43.515 27.9582 43.515Z" />
+                <path d="M27.9582 58.9199C31.5784 58.9199 34.5132 55.9851 34.5132 52.3648C34.5132 48.7446 31.5784 45.8098 27.9582 45.8098C24.3379 45.8098 21.4031 48.7446 21.4031 52.3648C21.4031 55.9851 24.3379 58.9199 27.9582 58.9199Z" />
+                <path d="M63.8902 31.8505H57.789V28.2161H73.7887V31.8505H67.671V48.5767H63.8902V31.8505Z" />
+                <path d="M86.5493 44.2141C86.229 45.3779 84.3972 48.8678 79.7142 48.8678C75.3515 48.8678 72.1508 45.6689 72.1508 41.1598C72.1508 36.6507 75.3515 33.4518 79.7142 33.4518C83.786 33.4518 86.9866 36.6507 86.9866 40.8688C86.9796 41.2192 86.9502 41.5689 86.8988 41.9156L86.8402 42.3237H75.787C76.1073 44.3879 77.6481 45.8135 79.7142 45.8135C81.4307 45.8135 82.4775 44.8253 82.7685 44.2141H86.5493ZM83.3504 39.8513C83.0302 37.9591 81.7784 36.5061 79.7142 36.5061C77.5036 36.5061 76.2226 37.9591 75.787 39.8513H83.3504Z" />
+                <path d="M89.1662 28.2161H92.8024V48.5767H89.1662V28.2161Z" />
+                <path d="M109.38 44.2141C109.06 45.3779 107.228 48.8678 102.543 48.8678C98.1807 48.8678 94.9819 45.6689 94.9819 41.1598C94.9819 36.6507 98.1807 33.4518 102.543 33.4518C106.617 33.4518 109.816 36.6507 109.816 40.8688C109.809 41.2192 109.779 41.5689 109.728 41.9156L109.671 42.3237H98.6163C98.9365 44.3879 100.479 45.8135 102.543 45.8135C104.26 45.8135 105.307 44.8253 105.598 44.2141H109.38ZM106.18 39.8513C105.859 37.9591 104.61 36.5061 102.543 36.5061C100.333 36.5061 99.0536 37.9591 98.6163 39.8513H106.18Z" />
+                <path d="M113.159 37.2326H110.687V33.7428H113.159V31.5596C113.159 29.349 114.585 27.9234 116.795 27.9234H119.996V31.1241H117.813C117.678 31.1206 117.544 31.1445 117.419 31.1943C117.294 31.2442 117.18 31.319 117.085 31.4141C116.99 31.5093 116.915 31.6229 116.865 31.7479C116.815 31.873 116.79 32.0069 116.794 32.1415V33.7428H119.994V37.2326H116.794V48.5786H113.157L113.159 37.2326Z" />
+                <path d="M135.666 41.169C135.666 45.6781 132.466 48.8769 128.103 48.8769C123.74 48.8769 120.54 45.669 120.54 41.169C120.54 36.669 123.74 33.461 128.103 33.461C132.466 33.461 135.666 36.6507 135.666 41.169ZM132.03 41.169C132.03 38.5503 130.284 36.8063 128.103 36.8063C125.922 36.8063 124.176 38.5503 124.176 41.169C124.176 43.7877 125.922 45.5317 128.103 45.5317C130.284 45.5317 132.03 43.7786 132.03 41.169Z" />
+                <path d="M137.846 33.7428H141.191L141.482 35.2068H141.627C141.915 34.8717 142.248 34.5779 142.617 34.3339C143.536 33.7459 144.609 33.4422 145.7 33.461C149.046 33.461 151.518 35.9333 151.518 39.7141V48.5859H147.882V39.9959C147.882 38.1055 146.573 36.7971 144.681 36.7971C142.789 36.7971 141.482 38.1055 141.482 39.9959V48.5768H137.846V33.7428Z" />
+                <path d="M174.66 43.0502C174.195 45.3779 172.363 48.8678 167.679 48.8678C163.316 48.8678 160.117 45.6689 160.117 41.1598C160.117 36.6507 163.316 33.4518 167.679 33.4518C172.363 33.4518 174.195 36.9416 174.66 39.1248H171.026C170.705 38.2812 169.862 36.7989 167.679 36.7989C165.495 36.7989 163.746 38.5411 163.746 41.1689C163.746 43.7968 165.492 45.5317 167.673 45.5317C169.854 45.5317 170.7 44.0786 171.02 43.0593L174.66 43.0502Z" />
+                <path d="M185.655 47.1238H185.51C185.217 47.465 184.874 47.7589 184.491 47.9949C183.532 48.5893 182.42 48.8925 181.292 48.8678C177.917 48.8678 175.91 46.8035 175.91 44.3605C175.91 41.4508 177.947 39.4195 182.018 39.4195H185.364V39.1285C185.364 37.4998 184.376 36.3652 182.745 36.3652C181.114 36.3652 180.273 37.4138 180.128 38.2574H176.492C176.812 35.8126 178.79 33.4573 182.745 33.4573C186.528 33.4573 189 35.9589 189 39.1303V48.5841H185.946L185.655 47.1238ZM185.364 42.1828H182.309C180.419 42.1828 179.546 42.9148 179.546 44.075C179.546 45.2352 180.39 45.9654 181.874 45.9654C184.085 45.9654 185.364 44.6844 185.364 42.6201V42.1828Z" />
+                <path d="M154.301 33.7428H157.938V48.5767H154.301V33.7428Z" />
+                <path d="M156.118 31.6108C157.307 31.6108 158.271 30.6473 158.271 29.4588C158.271 28.2702 157.307 27.3067 156.118 27.3067C154.93 27.3067 153.966 28.2702 153.966 29.4588C153.966 30.6473 154.93 31.6108 156.118 31.6108Z" />
+                <path d="M129.412 27.4879H133.194L129.702 31.706H126.648L129.412 27.4879Z" />
+            </svg>
+        );
+    }
+
+    return (
+        <svg
+            className={styles.svg}
+            style={applyCssVars(calcInlineVars(size))}
+            viewBox="0 0 72 72"
+            fill={color}
+            role="presentation"
+        >
+            <path d="M20.4652 19.2599C24.1268 19.2599 27.0951 16.2916 27.0951 12.63C27.0951 8.96833 24.1268 6 20.4652 6C16.8035 6 13.8352 8.96833 13.8352 12.63C13.8352 16.2916 16.8035 19.2599 20.4652 19.2599Z" />
+            <path d="M36.0447 19.2599C39.7063 19.2599 42.6746 16.2916 42.6746 12.63C42.6746 8.96833 39.7063 6 36.0447 6C32.383 6 29.4147 8.96833 29.4147 12.63C29.4147 16.2916 32.383 19.2599 36.0447 19.2599Z" />
+            <path d="M51.6241 19.2599C55.2858 19.2599 58.2541 16.2916 58.2541 12.63C58.2541 8.96833 55.2858 6 51.6241 6C47.9625 6 44.9942 8.96833 44.9942 12.63C44.9942 16.2916 47.9625 19.2599 51.6241 19.2599Z" />
+            <path d="M36.0447 34.8394C39.7063 34.8394 42.6746 31.8711 42.6746 28.2095C42.6746 24.5479 39.7063 21.5795 36.0447 21.5795C32.383 21.5795 29.4147 24.5479 29.4147 28.2095C29.4147 31.8711 32.383 34.8394 36.0447 34.8394Z" />
+            <path d="M36.0447 50.4201C39.7063 50.4201 42.6746 47.4518 42.6746 43.7902C42.6746 40.1286 39.7063 37.1602 36.0447 37.1602C32.383 37.1602 29.4147 40.1286 29.4147 43.7902C29.4147 47.4518 32.383 50.4201 36.0447 50.4201Z" />
+            <path d="M15.9028 60.2396H13.8352V59.0084H19.2508V60.2396H17.1832V65.9012H15.9028V60.2396Z" />
+            <path d="M23.5732 64.424C23.4652 64.8176 22.8448 65.9996 21.2596 65.9996C19.7824 65.9996 18.6988 64.9196 18.6988 63.3896C18.6988 61.8596 19.7788 60.7808 21.2596 60.7808C22.6372 60.7808 23.7208 61.8608 23.7208 63.2912C23.719 63.4118 23.709 63.532 23.6908 63.6512L23.6716 63.7892H19.93C20.038 64.4876 20.56 64.97 21.2596 64.97C21.8404 64.97 22.1944 64.6352 22.2928 64.4288L23.5732 64.424ZM22.4932 62.9468C22.3852 62.3072 21.9616 61.814 21.2632 61.814C20.5144 61.814 20.0812 62.3072 19.9336 62.9468H22.4932Z" />
+            <path d="M24.4587 59.0084H25.6899V65.9012H24.4587V59.0084Z" />
+            <path d="M31.3011 64.424C31.1931 64.8176 30.5727 65.9996 28.9875 65.9996C27.5103 65.9996 26.4279 64.9196 26.4279 63.3896C26.4279 61.8596 27.5079 60.7808 28.9875 60.7808C30.3663 60.7808 31.4499 61.8608 31.4499 63.2912C31.4475 63.4117 31.4375 63.532 31.4199 63.6512L31.3995 63.7892H27.6579C27.7659 64.4876 28.2879 64.97 28.9875 64.97C29.5683 64.97 29.9235 64.6352 30.0219 64.4288L31.3011 64.424ZM30.2211 62.9468C30.1119 62.3072 29.6895 61.814 28.9899 61.814C28.2411 61.814 27.8079 62.3072 27.6603 62.9468H30.2211Z" />
+            <path d="M32.5815 62.0612H31.7415V60.8792H32.5815V60.14C32.5815 59.3924 33.0615 58.91 33.8127 58.91H34.8927V59.99H34.1547C34.109 59.9888 34.0636 59.997 34.0212 60.014C33.9788 60.0309 33.9403 60.0564 33.9081 60.0887C33.8759 60.121 33.8506 60.1596 33.8338 60.2021C33.8169 60.2445 33.8089 60.2899 33.8103 60.3356V60.8768H34.8903V62.0588H33.8103V65.8987H32.5815V62.0612Z" />
+            <path d="M40.2002 63.3896C40.2002 64.916 39.1202 65.9996 37.6394 65.9996C36.1586 65.9996 35.0846 64.9196 35.0846 63.3896C35.0846 61.8596 36.1646 60.7808 37.6442 60.7808C39.1238 60.7808 40.2002 61.8644 40.2002 63.3896ZM38.969 63.3896C38.969 62.504 38.3786 61.9136 37.6394 61.9136C36.9002 61.9136 36.3098 62.504 36.3098 63.3896C36.3098 64.2752 36.9014 64.8668 37.6394 64.8668C38.3774 64.8668 38.969 64.2764 38.969 63.3896Z" />
+            <path d="M40.9382 60.8792H42.0698L42.1682 61.3712H42.2174C42.3154 61.2584 42.428 61.1591 42.5522 61.076C42.8637 60.8774 43.2268 60.7747 43.5962 60.7808C44.729 60.7808 45.5654 61.6208 45.5654 62.8976V65.8976H44.3354V62.9996C44.3354 62.7122 44.2213 62.4366 44.018 62.2334C43.8148 62.0302 43.5392 61.916 43.2518 61.916C42.9644 61.916 42.6888 62.0302 42.4856 62.2334C42.2824 62.4366 42.1682 62.7122 42.1682 62.9996V65.9048H40.9382V60.8792Z" />
+            <path d="M53.4001 64.0304C53.2429 64.8176 52.6225 65.9996 51.0373 65.9996C49.5602 65.9996 48.4766 64.9196 48.4766 63.3896C48.4766 61.8596 49.5566 60.7808 51.0373 60.7808C52.6225 60.7808 53.2429 61.9628 53.4001 62.7008H52.1689C52.0913 62.4646 51.9389 62.2601 51.7347 62.1181C51.5306 61.9761 51.2858 61.9043 51.0373 61.9136C50.2982 61.9136 49.7078 62.504 49.7078 63.3896C49.7078 64.2752 50.2982 64.8668 51.0373 64.8668C51.2939 64.8785 51.5466 64.8013 51.7527 64.6483C51.9589 64.4953 52.1059 64.2757 52.1689 64.0268L53.4001 64.0304Z" />
+            <path d="M57.1225 65.4092H57.0733C56.974 65.5247 56.8574 65.6243 56.7277 65.7044C56.4037 65.9048 56.0287 66.0073 55.6477 65.9996C54.5053 65.9996 53.8262 65.3 53.8262 64.4732C53.8262 63.4892 54.515 62.7992 55.8937 62.7992H57.0265V62.7008C57.0265 62.1488 56.6917 61.7648 56.1397 61.7648C55.5877 61.7648 55.2997 62.1248 55.2541 62.4056H54.0206C54.1286 61.5788 54.7981 60.7808 56.1373 60.7808C57.4177 60.7808 58.2541 61.628 58.2541 62.7008V65.9012H57.2209L57.1225 65.4092ZM57.0241 63.7292H55.9897C55.3501 63.7292 55.0538 63.9692 55.0538 64.3688C55.0538 64.7684 55.3393 65.0084 55.8421 65.0084C56.5897 65.0084 57.0241 64.5752 57.0241 63.8768V63.7292Z" />
+            <path d="M46.5086 60.8792H47.7386V65.9012H46.5086V60.8792Z" />
+            <path d="M47.1242 60.1569C47.5265 60.1569 47.8526 59.8308 47.8526 59.4285C47.8526 59.0262 47.5265 58.7001 47.1242 58.7001C46.7219 58.7001 46.3958 59.0262 46.3958 59.4285C46.3958 59.8308 46.7219 60.1569 47.1242 60.1569Z" />
+            <path d="M38.0834 58.7624H39.3626L38.1806 60.1904H37.1475L38.0834 58.7624Z" />
+        </svg>
+    );
+};
+
+export default TelefonicaLogoImage;
