+++ src/logo-vivo.tsx
@@ -1,0 +1,64 @@
+import * as React from 'react';
+import * as styles from './logo.css';
+import {getVivoSkin} from './skins/vivo';
+import {applyCssVars} from './utils/css';
+import {calcInlineVars} from './logo-common';
+
+import type {LogoImageProps} from './logo-common';
+
+const VivoLogoImage = ({size, type, isInverse, isDarkMode}: LogoImageProps): JSX.Element => {
+    const {colors} = getVivoSkin();
+    const color = isInverse && !isDarkMode ? colors.inverse : colors.brand;
+
+    if (type === 'vertical') {
+        return (
+            <svg
+                className={styles.svg}
+                style={applyCssVars(calcInlineVars(size))}
+                viewBox="0 0 72 72"
+                fill={color}
+                role="presentation"
+            >
+                <path d="M24.7302 55.7681V54.5346C24.7302 54.3255 24.6222 53.5308 23.8252 53.0943L14.7817 47.9903C13.8238 47.4665 13.0106 48.0844 12.8085 48.6288L12.1998 49.9864C11.8736 50.7169 12.2181 51.5943 12.9831 51.9228L19.4217 55.1272V55.1754L12.9831 58.3752C12.2204 58.706 11.8736 59.5835 12.1998 60.3117L12.8085 61.6692C13.0106 62.2159 13.8238 62.8315 14.7817 62.3101L23.8252 57.206C24.6222 56.7719 24.7302 55.9771 24.7302 55.7681Z" />
+                <path d="M24.7302 33.3465V32.1152C24.7302 31.9062 24.6222 31.1114 23.8252 30.675L14.7817 25.5709C13.8238 25.0472 13.0106 25.6651 12.8085 26.2118L12.1998 27.5694C11.8736 28.2998 12.2181 29.1773 12.9831 29.5058L19.4217 32.7102V32.7584L12.9831 35.9605C12.2204 36.2913 11.8736 37.1665 12.1998 37.8969L12.8085 39.2545C13.0106 39.8012 13.8238 40.4168 14.7817 39.8954L23.8252 34.7913C24.6222 34.3503 24.7302 33.5555 24.7302 33.3465Z" />
+                <path d="M12.1216 18.3697C12.1216 21.8865 14.9746 24.7372 18.4914 24.7372C22.0105 24.7372 24.8611 21.8865 24.8611 18.3697C24.8611 14.8506 22.0105 12 18.4914 12C14.9746 12 12.1216 14.8506 12.1216 18.3697ZM21.0043 18.3697C21.0043 19.7572 19.8788 20.8827 18.4914 20.8827C17.1016 20.8827 15.9761 19.7572 15.9761 18.3697C15.9761 16.9823 17.1016 15.8545 18.4914 15.8545C19.8788 15.8568 21.0043 16.9823 21.0043 18.3697Z" />
+                <path d="M8.4969 41.4482C9.87513 41.4482 10.9915 42.5669 10.9915 43.9451C10.9915 45.3234 9.87283 46.442 8.4969 46.442C7.11637 46.442 6 45.3234 6 43.9451C6 42.5669 7.11637 41.4482 8.4969 41.4482Z" />
+                <path d="M13.7939 46.1158C12.9417 46.1158 12.2526 45.4267 12.2526 44.5768V43.2882C12.2526 42.4383 12.9417 41.7491 13.7939 41.7491H23.1935C24.0434 41.7491 24.7325 42.4383 24.7325 43.2882V44.5768C24.7325 45.429 24.0411 46.1158 23.1935 46.1158H13.7939Z" />
+                <path d="M57.9111 22.8788C57.9111 24.5074 57.3346 27.048 55.5406 28.842C54.9089 29.4737 54.3645 29.8504 53.9258 30.1329C53.0827 30.6773 52.4717 31.0586 52.8117 31.9683C53.1448 32.8641 54.1876 32.5862 55.1731 32.3634C55.49 32.2875 61.676 30.8082 61.7541 30.8036C63.4149 30.489 65.0205 31.541 65.411 33.1673C65.4133 33.1742 65.9072 35.2393 65.9118 35.2576C66.3046 36.884 65.3513 38.5493 63.7296 39.0248C63.6607 39.0547 56.9004 40.681 56.889 40.6833C55.6049 41.0095 54.8538 41.6481 54.8538 42.8862C54.8538 43.4398 55.1731 44.0416 55.5681 44.54C55.5681 44.54 62.9118 53.6433 62.9509 53.7122C63.9501 55.0766 63.6951 56.9763 62.3973 58.0307C62.3927 58.0352 60.7434 59.3721 60.7273 59.3836C59.4318 60.438 57.5184 60.2978 56.3882 59.0391C56.3308 58.9908 51.7688 53.3837 51.2841 52.7888C50.7995 52.1938 50.1907 51.2888 49.3454 51.2888C48.5024 51.2888 47.8937 52.1938 47.409 52.7888C46.9266 53.3837 42.3624 58.9908 42.3049 59.0391C41.1748 60.2978 39.2613 60.438 37.9635 59.3836C37.9497 59.3721 36.2935 58.0307 36.2935 58.0307C34.9957 56.9763 34.7407 55.0743 35.74 53.7122C35.7767 53.641 43.1227 44.54 43.1227 44.54C43.5155 44.0416 43.8371 43.4398 43.8371 42.8862C43.8371 41.6481 43.0859 41.0118 41.8019 40.6833C41.7904 40.681 35.0302 39.0547 34.9613 39.0248C33.3395 38.5493 32.3863 36.8817 32.7768 35.2576C32.7814 35.2393 33.2775 33.1673 33.2775 33.1673C33.668 31.541 35.2714 30.4912 36.9344 30.8036C37.0125 30.8059 43.1985 32.2875 43.5155 32.3634C44.5032 32.5862 45.5438 32.8641 45.8769 31.9683C46.2168 31.0563 45.6058 30.6773 44.7651 30.1329C44.3264 29.8504 43.7797 29.4714 43.1503 28.842C41.3563 27.048 40.7774 24.5074 40.7774 22.8788C40.7774 18.2066 44.6732 14.421 49.3431 14.421C54.0153 14.4187 57.9111 18.2066 57.9111 22.8788Z" />
+            </svg>
+        );
+    }
+
+    if (type === 'imagotype') {
+        return (
+            <svg
+                className={styles.svg}
+                style={applyCssVars(calcInlineVars(size))}
+                viewBox="0 0 195 72"
+                fill={color}
+                role="presentation"
+            >
+                <path d="M23.3761 61.4297H26.5561C27.0861 61.4297 29.1472 61.1549 30.266 59.0938L43.4176 35.7743C44.7721 33.301 43.1821 31.2006 41.7688 30.6903L38.2748 29.1199C36.3904 28.2759 34.133 29.1592 33.2693 31.1417L25.0054 47.7481H24.8876L16.6433 31.1417C15.7992 29.1788 13.5222 28.2759 11.6575 29.1199L8.16346 30.6903C6.75015 31.2006 5.16018 33.301 6.5146 35.7743L19.6858 59.0938C20.8047 61.1549 22.8462 61.4297 23.3761 61.4297Z" />
+                <path d="M81.204 61.4297H84.3839C84.9139 61.4297 86.975 61.1549 88.0939 59.0938L101.265 35.7743C102.62 33.301 101.03 31.2006 99.6163 30.6903L96.1223 29.1199C94.2378 28.2759 91.9805 29.1592 91.1364 31.1417L82.8725 47.7481H82.7547L74.4908 31.1417C73.6467 29.1788 71.3697 28.2759 69.5049 29.1199L66.0109 30.6903C64.5976 31.2006 63.0077 33.301 64.3621 35.7743L77.5137 59.0938C78.6129 61.1549 80.674 61.4297 81.204 61.4297Z" />
+                <path d="M119.815 28.904C110.746 28.904 103.385 36.265 103.385 45.3337C103.385 54.4024 110.746 61.7634 119.815 61.7634C128.884 61.7634 136.244 54.4024 136.244 45.3337C136.244 36.265 128.903 28.904 119.815 28.904ZM119.815 51.8114C116.242 51.8114 113.337 48.9062 113.337 45.3337C113.337 41.7612 116.242 38.8561 119.815 38.8561C123.387 38.8561 126.292 41.7612 126.292 45.3337C126.312 48.9062 123.407 51.8114 119.815 51.8114Z" />
+                <path d="M60.3184 19.5603C60.3184 23.1132 57.4329 25.9987 53.88 25.9987C50.3271 25.9987 47.4416 23.1132 47.4416 19.5603C47.4416 16.0074 50.3271 13.1219 53.88 13.1219C57.4329 13.1219 60.3184 16.0074 60.3184 19.5603Z" />
+                <path d="M48.2661 33.2226C48.2661 31.0241 50.0327 29.2574 52.2312 29.2574H55.5485C57.747 29.2574 59.5136 31.0241 59.5136 33.2226V57.4647C59.5136 59.6632 57.7274 61.4298 55.5485 61.4298H52.2312C50.0327 61.4298 48.2661 59.6632 48.2661 57.4647V33.2226Z" />
+                <path d="M178.919 19.5409C178.919 21.5627 178.192 24.7427 175.974 26.9608C175.189 27.746 174.502 28.2171 173.952 28.5704C172.912 29.2574 172.146 29.7285 172.559 30.867C172.971 31.9859 174.266 31.6326 175.503 31.3577C175.896 31.2596 183.61 29.4144 183.708 29.4144C185.769 29.0219 187.771 30.337 188.262 32.3588C188.262 32.3588 188.871 34.9499 188.89 34.9695C189.381 36.9913 188.184 39.0721 186.162 39.6609C186.083 39.7002 177.662 41.722 177.643 41.722C176.053 42.1342 175.11 42.9194 175.11 44.4701C175.11 45.1571 175.503 45.903 175.994 46.5312C175.994 46.5312 185.141 57.8769 185.18 57.9554C186.417 59.6632 186.103 62.0187 184.493 63.3338C184.493 63.3338 182.432 65.0023 182.413 65.0219C180.803 66.3371 178.408 66.1604 177.014 64.5901C176.936 64.5312 171.263 57.5432 170.655 56.7973C170.046 56.0514 169.3 54.9325 168.24 54.9325C167.18 54.9325 166.434 56.0514 165.826 56.7973C165.217 57.5432 159.544 64.5116 159.466 64.5901C158.053 66.1604 155.677 66.3371 154.068 65.0219C154.048 65.0023 151.987 63.3338 151.987 63.3338C150.378 62.0187 150.044 59.6435 151.3 57.9554C151.339 57.8769 160.487 46.5312 160.487 46.5312C160.977 45.903 161.37 45.1571 161.37 44.4701C161.37 42.9194 160.428 42.1342 158.838 41.722C158.818 41.722 150.397 39.7002 150.319 39.6609C148.297 39.0721 147.119 36.9913 147.59 34.9695C147.59 34.9499 148.218 32.3785 148.218 32.3588C148.709 30.337 150.711 29.0219 152.772 29.4144C152.87 29.4144 160.565 31.2596 160.977 31.3577C162.214 31.6326 163.51 31.9859 163.922 30.867C164.354 29.7285 163.588 29.2574 162.528 28.5704C161.978 28.2171 161.311 27.746 160.526 26.9608C158.288 24.723 157.562 21.5627 157.562 19.5409C157.562 13.7307 162.41 9 168.24 9C174.07 9 178.919 13.711 178.919 19.5409Z" />
+            </svg>
+        );
+    }
+
+    return (
+        <svg
+            className={styles.svg}
+            style={applyCssVars(calcInlineVars(size))}
+            viewBox="0 0 72 72"
+            fill={color}
+            role="presentation"
+        >
+            <path d="M46.8016 17.1117C46.8016 19.2508 46.0737 22.5899 43.8127 24.9457C43.0178 25.7738 42.3281 26.2722 41.777 26.6429C40.714 27.3592 39.9446 27.8575 40.3732 29.0552C40.7938 30.2313 42.1074 29.8665 43.3494 29.574C43.7492 29.4753 51.549 27.5301 51.6472 27.5253C53.7419 27.1136 55.7661 28.4943 56.2571 30.6297C56.2594 30.6381 56.8833 33.3526 56.8891 33.3766C57.3836 35.5121 56.182 37.7017 54.137 38.3264C54.0492 38.3673 45.5261 40.5028 45.5111 40.5064C43.8924 40.9361 42.945 41.7727 42.945 43.399C42.945 44.126 43.3483 44.9157 43.8462 45.5705C43.8462 45.5705 53.1064 57.5285 53.1538 57.6212C54.4143 59.4123 54.0931 61.9101 52.4548 63.2944C52.449 63.3004 50.3694 65.0579 50.3486 65.0735C48.7137 66.459 46.3014 66.276 44.8779 64.6221C44.804 64.5571 39.0538 57.1926 38.4426 56.4114C37.8326 55.6314 37.0654 54.4421 35.999 54.4421C34.9326 54.4421 34.1689 55.6302 33.5566 56.4114C32.9477 57.1926 27.194 64.5583 27.1224 64.6221C25.6978 66.2748 23.2854 66.459 21.6494 65.0735C21.6321 65.0579 19.5501 63.3004 19.5444 63.2944C17.9084 61.9101 17.586 59.4123 18.8454 57.6212C18.8927 57.5285 28.153 45.5705 28.153 45.5705C28.6498 44.9157 29.0541 44.126 29.0541 43.399C29.0541 41.7727 28.1079 40.9361 26.4881 40.5064C26.4731 40.5028 17.9511 38.3673 17.8633 38.3264C15.8183 37.7017 14.6179 35.5121 15.1101 33.3766C15.1147 33.3526 15.7409 30.6381 15.7421 30.6297C16.2331 28.4943 18.2561 27.1124 20.352 27.5253C20.4502 27.5301 28.2489 29.4753 28.6486 29.574C29.8929 29.8665 31.2054 30.2313 31.626 29.0552C32.0546 27.8587 31.2851 27.3592 30.2234 26.6429C29.6711 26.2722 28.9814 25.775 28.1865 24.9457C25.9243 22.5887 25.1964 19.2508 25.1964 17.1117C25.1987 10.9751 30.1101 6 36.0002 6C41.8902 6 46.8016 10.9751 46.8016 17.1117Z" />
+        </svg>
+    );
+};
+
+export default VivoLogoImage;
